Revert "flake.lock: Update"
This reverts commit 173b21361ca863d47c81d477dc3634393b04a80d.

In this update the maintainers removed the 6.10 kernel series because it
reached end-of-life.

My experiments showed that my Thinkpad X1 needs at least 6.9, thus I
don't want to go back to the stable 6.6 series, which is the default in
NixOS 24.04.

I tried to switch to the 6.11 series but a compatbile ZFS package is not
in nixpkgs yet: the zfs-unstable is only 2.2.6 and 6.11 support starts
from ZFS 2.3.0.

Let's wait a bit with the update...